requirements.txt
Django>=1.10,<1.11
wagtail>=1.7,<1.8
psycopg2==2.6
wagtail-embed-videos
django-simple-captcha
